fix!: Remove non working all_project flag from snapshot creator
The `all_project` flag from snapshot creator intended to create
snapshots in the volume owner's project, but even as an admin user
the snapshots created using the `all_project` flag were created
in the admin user project (the one linked to the token).

This commit remove this non working flag.

Signed-off-by: Nicolas Belouin <nicolas.belouin@gandi.net>