feat: creator: immediately delete snapshots in error
When a snapshot isn't completed (status is error) we should try to clean
it right away as it's unusable and it's size is taken in account in the
volumes quotas.

The delete call is retried in case of intermittent network failures and
such then we wait for the snapshot to be removed.